+++ custom_components/huawei_fusion_hub/__init__.py
@@ -13,6 +13,8 @@ PLATFORMS = [
     Platform.BINARY_SENSOR,
     Platform.SWITCH,
     Platform.SELECT,
+    Platform.NUMBER,
+    Platform.BUTTON,
 ]